Catch runtime error for `edit student`
