+++ src/main/java/tfifteenfour/clipboard/logic/commands/Command.java
@@ -2,6 +2,7 @@ package tfifteenfour.clipboard.logic.commands;
 
 import tfifteenfour.clipboard.logic.CurrentSelection;
 import tfifteenfour.clipboard.logic.commands.exceptions.CommandException;
+import tfifteenfour.clipboard.logic.parser.exceptions.ParseException;
 import tfifteenfour.clipboard.model.Model;
 
 /**
@@ -23,7 +24,7 @@ public abstract class Command {
      * @return feedback message of the operation result for display
      * @throws CommandException If an error occurs during command execution.
      */
-    public abstract CommandResult execute(Model model, CurrentSelection currentSelection) throws CommandException;
+    public abstract CommandResult execute(Model model, CurrentSelection currentSelection) throws CommandException, ParseException;
 
     public boolean getWillModifyState() {
         return this.willModifyState;
